group-permission-settings: Add function to get list of setting names.
This commit adds a function which returns type checked list of
group permission setting names so that all the type checking
has to be done only once.
