+++ tools/test-js-with-node
@@ -111,6 +111,7 @@ EXEMPT_FILES = make_set(
         "web/src/giphy.ts",
         "web/src/giphy_state.ts",
         "web/src/global.ts",
+        "web/src/group_permission_settings.ts",
         "web/src/group_setting_pill.ts",
         "web/src/hash_util.ts",
         "web/src/hashchange.ts",
